Add PlayFab Explorer to extension
This commit adds an explorer view to the extension. A PlayFab icon is
added to the activity bar and a tree view of Studio/Title is displayed
once the developer has signed in.

Also contains some clean up and fixes to the login code.

Details

Add viewsContainers, views and menus items to package.json along with
refresh and createTitle commands. createTitle currently just pops
a toast (and is bound to the tree node, rather than a menu, which needs
to be fixed).

Add SVG file for PlayFab logo and refresh icon

Add a PlayFabExplorer and a PlayFabStudioTreeProvider. The latter gets
data by calling the GetStudio developer API. It also indicates that the
developer needs to sign-in if necessary.

Add Studio and Title models to support the API call and the tree data.

I'm not entirely convinced that the data refresh and event firing code
is quite right sequence wise. It works correctly, but I suspect it
could be cleaner.

Cleanup

Add explicit function return types.

Be sure to await all async function calls.