Add SignIn items to tree
Update the PlayFabStudioTreeViewProvider to return SignIn and
CreateAccount tree view items in the case where the user is not
logged in.

Details

Amend getChildren to check for sign-in state in the case of a null or
undefined entry. Such cases mean we are being asked for the
'root nodes'. If the user is signed in, we return the Studio nodes as
before. If the user is not signed in, we return two 'Command' items; one
for sign-in and one for account creation.

Add localizable strings for the command titles.

Add an await to the HTTP API call in updateStudioData and propogate the
Promise signature appropriately.

Remove the code in updateStudioData that displayed a popup message if
the user was not signed in.

Add an isLoggedIn method and use where appropriate.

Some cleanup of != => !==