Make HTTP Client Mockable
Add an IHttpClient interface and amend PlayFabHttpHelper to implement
it.

Amend PlayFabLoginManager, PlayFabExplorer and PlayFabStudioTreeProvider
to take an optional IHttpClient parameter in their constructors. If one
is not provided, just create a PlayFabHttpHelper. This will allow us to
pass in a mock IHttpClient for testing purposes.

Amend all HTTP call sites to use the private field rather than consing
up one in each function.