Better 2FA login flow
This PR changes the login flow to only ask for a 2FA code if the account
actually requires it.

Details

Add a getUserInputForTwoFA method to IPlayFabLoginInputGatherer

In the login flow, first ask for username/password, then make the login
call. Iff it fails with error 1246, ask for TwoFA code and make login
call again.

Coin an ErrorResponse type for use in errorcallback when making HTTP API
calls. Amend various callsites and implementations of errorcallback
accordingly.

Amend various showError methods to also take ErrorResponse.

Update various mocks to new HTTP API signature.

Add test for logging in with an account that requires TWO FA