Add support for RegisterFunction
This commit adds rudimentary support for registering an Azure function
with PlayFab for a given title, such that said function can then be
called by a game or, in future, in response to a playstram event.

Details

Add commands for registering and unregistering functions.
Add methods to the input gatherer for PlayFab Explorer to get input for
register and unregister function.

The register command makes an API call to convert the title secret into
an entity tokem and then calls the RegisterFunction API.

Add URI paths for RegisterFunction, UnregisterFunction and
GetEntityToken

Also rename GetAdminBaseUrl to GetPlayFabBaseUrl.

Add GetLastPathPartFromUri function which is used to 'suggest' a
function name as part of gathering input for the register call.

 GetLastPathPartFromUri returns the part of the URI after the last /
 character that occurs after the host name.

Add various tests for GetLastPathPartFromUri.

Add menu commands and NLS strings for RegisterFunction and
UnregisterFunction.