Fix palette commands
This PR changs the various palette commands that operate at the Studio
or Title level to present a picker to allow the developer to choose
a Studio and/or Title to have the command operator on.

Prior to this change, while the commands worked correctly when invoked
from a tree node menu item, they did not work when invoked from the
command palette.

Details

Factor PlayFabStudioTreeProvider out into its own file. Also add an api
file for the tree provider.

Make interface naming consistent.