Improve title data experience
This PR improves the experience around title data by opening title data
in a text window, thus allowing direct editing and then subsequent
upload via the Set Title Data menu.

Details

Amend PlayFabExplorer.getTitleData to open a JSON text window and
populate it with the data returned by the PlayFab API. Also amend the
processing of user input such that if no keys are specified, we request
all the key/value pairs via the PlayFab API.

Amend PlayFabExplorer.setTitleData to check whether the active text
windows is JSON and if so, parse it and call the PlayFab SetTitleData
API once for each field in the resultant object. If the active text
buffer is not JSON or there is no active text buffer, use the previous
approach of asking the user for a key and value.

Refactor some of the code related to the above into two seperate
methods; setTitleDataFromJsonFile and setTitleDataFromUserInput.

Also move the code that calls the PlayFab SetTitleData API into a helper
method; makeSetTitleDataApiCall

Update version number to 0.0.6