Local debugging support
This PR adds support for enabling and disabling local CloudScript
debugging for Azure Functions based CloudScript.

Details

Add Enable Local Debugging and Disable Local Debugging menu items to
Title nodes in the explorer. Note: Local debugging is an 'all or
nothing' affair, it's either on or off.

When handling the Enable Local Debugging command, write a file to the
TEMP directory. The file is called playfab.local.settings.json and
contains a LocalApiServer property with the value
http://localhost:7071/.

When handling the Disable Local Debugging command, delete the above
file.

Both commands pop a toast when complete indicating success or failure.