Improve title data support
This PR improves the experience around title data by opening title data
in a JSON text window, thus allowing direct editing and then subsequent
upload via the Set Title Data menu.

Details

Amend PlayFabExplorer.getTitleData to open a JSON text window and
populate it with the data returned by the PlayFab API. Also amend the
processing of user input such that if no keys are specified, we request
all the key/value pairs via the PlayFab API.

Add a setting to control the number of spaces used to control indenting
when we populate the JSON text window.

Amend PlayFabExplorer.setTitleData to check whether the active text
window is JSON and if so, parse it and call the PlayFab SetTitleData
API once for each field in the resultant object. If the active text
buffer is not JSON or there is no active text buffer, use the previous
approach of asking the user for a key and value.

Refactor some of the code related to the above into two seperate
methods; setTitleDataFromJsonFile and setTitleDataFromUserInput.

Also move the code that calls the PlayFab SetTitleData API into a helper
method; makeSetTitleDataApiCall

Add EscapeValue and UnescapeValue helpers to PlayFabDataHelpers. These
two functions should roundtrip correctly in both directions. Tests added
to check such roundtripping.

getTitleData uses UnescapeValue to provide actual JSON in the resulting
editor rather than the escaped strings that PlayFab stores.

setTitleData uses EscapeValue to ensure we are passing strings
(or null) when making the API call to PlayFab.

Update version number to 0.0.10