Add List and Unregister function support
This PR adds support for calling the ListFunctions and
UnregisterFunction APIs.

Details

Add a List Functions menu to title nodes.

Add a getUserInputForListFunctions method. This solicits input from the
user using two pick lists, one for the trigger type (how the function is
called) and one for the invocation source (what causes the function to
be called). Note that these are currently ignored by MainServer.

Add a ListFunctions method that makes the API call. Generate a markdown
file that shows a table of the results.

Implement the UnregisterFunction method.

Make a helper method for retrieving the title entity token.

Add initial tests for ListFunctions, RegisterFunction and
UnregisterFunction