Fix palette commands
This PR changs the various palette commands that operate at the Studio
or Title level to present a picker to allow the developer to choose
a Studio and/or Title to have the command operate on.

Prior to this change, while the commands worked correctly when invoked
from a tree node menu item, they did not work when invoked from the
command palette.

Details

Factor PlayFabStudioTreeProvider out into its own file. Also add an API
file for the tree provider. Move the definition of TreeNode into the API
file.

Add getStudioTreeNodeFromUser and getTitleTreeNodeFromUser to
PlayFabExplorer to ask the user to pick a Studio or Title respectively.

Add a getTitleFromTreeNode method to PlayFabExplorer to get a title from
a title tree node. Add a similar getStudioFromTreeNode method for
getting a studio from a studio tree node. Both these methods take a tree
node and, if it is null, call the appropriate method to ask the user to
pick a node.

Amend the various command lambdas to take a treeNode and call
getStudioFromTreeNode or getTitleFromTreeNode as appropriate.

Make interface naming consistent.

Update readme with additional pallete commands.

Remove unused string from package.nls.json