PR feedback and bugs
Address PR feedback around URI construction.
Update Open in Game Manage to respect private cloud.
Fix issue where private cloud was not honoured on first run.

Details

Add a call to updateCloud in the activate method. This ensures that if
a private cloud is set it will be honoured when the extension is first
activated.

Add a GetGameManagerUrl method to PlayFabUriHelpers and use that method
as part of the Open in Game Manager command.

Use ternery operator for Uri construction in PlayFabUriHelpers per PR
feedback.