Add support for private clouds (#35)
* Add support for private clouds

This PR adds a cloud name setting to the extension. This setting is then
used to construct based URIs for API calls.

Details

Add setting definition to package.json.

Add localizable string description for new setting to package.nls.json.

Rename PlayFabUriConstants to PlayFabUriHelpers.

Update PlayFabUriHelpers.GetPlayFabBaseUri to account for the value of
the cloud field.

Add a GetPlayFabBaseEditorUri method to PlayFabUriHelpers to handle the
cloud field for such URIs.

Make PlayFabUriHelpers.editorBaseUrl private.

Update various call sites to use the above helpers.

Add a configuration change handler to activate method in order to update
the cloud field of PlayFabUriHelpers whenever the setting is changed.

Update readme.md with new setting description

* PR feedback and bugs

Address PR feedback around URI construction.
Update Open in Game Manage to respect private cloud.
Fix issue where private cloud was not honoured on first run.

Details

Add a call to updateCloud in the activate method. This ensures that if
a private cloud is set it will be honoured when the extension is first
activated.

Add a GetGameManagerUrl method to PlayFabUriHelpers and use that method
as part of the Open in Game Manager command.

Use ternery operator for Uri construction in PlayFabUriHelpers per PR
feedback.