Sign-out on cloud change
If the user changes the cloud setting, then if the user is signed in, we
need to sign them out of the previous cloud.

Details

Add code to updateCloud to check if the cloud is actually changing. If
it is, then check to see if we are signed in by checking for a session
count that is greater than zero. If we are signed in, run the sign-out
command. In either case, store the new cloud value. NOTE: We sign-out
*before* storing the new cloud value because the cloud value affects the
base URI for API calls.

Remove the cloud property from IPlayFabSession as it was not used.