Add timeouts to API calls (#46)
This PR adds timeout support to API calls to PlayFab. By default the
timeout is 5000ms.

Details

Add a timeoutMilliseconds property to IHttpClient. Set this to 5000 in
PlayFabHttpClient. Amend makeApiCall, makeEntityApiCall and
makeTitleApiCall to use the asyncOr and timeout functions from
PlayFabPromiseHelpers.

Add a LoginWhenLoggedOutTimeout test that verifies that timeout fires
when Login takes too long.