Add title name to function list output
This PR adds the name of the title to the header comment in the output
of List functions, List HTTP functions and List Queued functions.

Details

Add an IPlayFabConfig interface and PlayFabConfig class to abstract
reading the config settings.

Update functionListHeader, functionListHTTPHeader and
functionListQueuedHeader to have a placeholder for the title name.

Plumb title object and the showTitleIds setting into
getMarkDownForFunctionList, getMarkDownForHttpFunctionList and
getMarkDownForQueuedFunctionList. If showTitleIds is true, append the
title id in parentheses, like we do in the tree view.

Update PlayFabStudioTreeProvider to use PlayFabConfig/IPlayFabConfig.

Some minor re-ordering of imports.